refactor(parser) rename to input
